Create a GroupItem datasource for resolving VolunteerGroup requests
