+++ src/data/groups/index.js
@@ -1,5 +1,3 @@
-import schema from './schema';
-import resolver from './resolver';
 import dataSource from './data-source';
 
-export { schema, resolver, dataSource };
+export { dataSource };
